work around basic auth dialog box
When the user is anonymous and visits the home page, an ajax call was
being made to fetch mapmeter data. The response was a basic auth
challenge, and the browser would throw up a basic auth dialog box. This
conditionally only makes that request if there is no login link in the
header.